requirements.txt
python_chargepoint
google-api-python-client
google-auth-httplib2
google-auth-oauthlib